requirements.txt
# Core dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.25.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
httpx>=0.27.0

# Performance and Monitoring
psutil>=5.9.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.1
structlog>=23.2.0
prometheus-client>=0.20.0

# Web scraping and content extraction (internet_agent_advanced)
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
lxml>=5.2.1
trafilatura>=1.7.0
tldextract>=5.1.2
aiolimiter>=1.1.0
cachetools>=5.3.3
sqlitedict>=2.1.0
playwright>=1.45.0
PyGithub>=2.3.0
huggingface_hub>=0.23.0
user-agents>=2.2.0
robotexclusionrulesparser>=1.7.1
simhash>=2.0.1
sentence-transformers>=2.5.1
boto3>=1.34.0

# Feedback Loop + Oracle Meta-Loop Dependencies
asyncio-mqtt>=0.11.0  # For advanced event bus (if using MQTT)
aiofiles>=23.0.0  # Async file operations
watchfiles>=0.20.0  # File watching for hot reloading
croniter>=1.4.0  # Cron job scheduling for Oracle ticks

# Vector Memory System Dependencies
qdrant-client>=1.11.3
faiss-cpu>=1.8.0
numpy>=1.26.4
tenacity>=8.4.1

# AI/ML Dependencies (Optional - for LoRA and model management)
openai>=1.0.0  # OpenAI API client
torch>=2.0.0  # PyTorch for model operations
transformers>=4.30.0  # Hugging Face transformers
peft>=0.4.0  # Parameter-Efficient Fine-Tuning (LoRA)
accelerate>=0.20.0  # Model acceleration
safetensors>=0.3.0  # Safe tensor serialization

# Data Processing
pandas>=2.0.0  # Data analysis for feedback processing
scipy>=1.10.0  # Scientific computing for statistical analysis

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.5.0
detect-secrets>=1.4.0
bandit>=1.7.0
safety>=2.3.0

# Additional dependencies
pyyaml>=6.0.0
click>=8.0.0  # CLI interfaces
rich>=13.0.0  # Rich text and beautiful formatting
